consolidate integration test files
